Switch CSES updater to use our task system.
The updater was started multiple times and ended up sending a lot of
requests to cses.fi, as reported in #294.

This should fix the issue, hopefully people using the bot will update so
that this won't stay an issue for too long. Thanks a lot to @ollpu for
reporting this, and I'm deeply sorry for the inconvenience caused.